Refactor _info, change behavior of mapping
What used to be _TMUX (_attr in an earlier commit) now has raw
information from tmux's FORMATS (see the official tmux(1) manual
for more details).

Instead of using the tmux object as a dictionary, the object will
set attributes for formats w/o the namespace.

So what used to be ``s.get('session_name')`` can now be retrieved
as ``s.name``.

Window and pane still have to be switched.