ci(docs[dorny/paths-filter]): 2.7.0 -> 2.11.1 to fix warning
The following actions uses node12 which is deprecated and will be forced to run on node16: dorny/paths-filter@v2.7.0. For more info: https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/

The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/